apis/speech_to_text_api/requirements.txt
# HTTP client
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0

# Configuration
python-dotenv==1.2.1

# FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.12.4

# Utilities
certifi==2025.11.12
idna==3.11
anyio==4.11.0
sniffio==1.3.1

